Added a refresh button to the list
-Took the opportunity to cleanup the access request list component.
